Cleaning up and moving some file contents for better readability
