+++ ukvm/ukvm_hv_kvm_x86_64.c
@@ -29,9 +29,11 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <sys/ioctl.h>
+#include <elf.h>
 #include <sys/mman.h>
 #include <string.h>
 #include <stdlib.h>
+#include <unistd.h>
 #include <linux/kvm.h>
 
 #include "ukvm.h"
@@ -241,4 +243,72 @@ void ukvm_hv_vcpu_loop(struct ukvm_hv *hv)
         }
         } /* switch(run->exit_reason) */
     }
+}
+
+void ukvm_hv_fill_elf_prstatus(x86_elf_prstatus *prstatus,
+	 struct ukvm_hv *hv, struct ukvm_dump_core *info)
+{
+    memset(prstatus, 0, sizeof(x86_elf_prstatus));
+    prstatus->regs.r8 = hv->b->kregs.r8;
+    prstatus->regs.r9 = hv->b->kregs.r9;
+    prstatus->regs.r10 = hv->b->kregs.r10;
+    prstatus->regs.r11 = hv->b->kregs.r11;
+    prstatus->regs.r12 = hv->b->kregs.r12;
+    prstatus->regs.r13 = hv->b->kregs.r13;
+    prstatus->regs.r14 = hv->b->kregs.r14;
+    prstatus->regs.r15 = hv->b->kregs.r15;
+    prstatus->regs.rbp = hv->b->kregs.rbp;
+    prstatus->regs.rsp = hv->b->kregs.rsp;
+    prstatus->regs.rdi = hv->b->kregs.rdi;
+    prstatus->regs.rsi = hv->b->kregs.rsi;
+    prstatus->regs.rdx = hv->b->kregs.rdx;
+    prstatus->regs.rcx = hv->b->kregs.rcx;
+    prstatus->regs.rbx = hv->b->kregs.rbx;
+    prstatus->regs.rax = hv->b->kregs.rax;
+    prstatus->regs.rip = hv->b->kregs.rip;
+    prstatus->regs.eflags = hv->b->kregs.rflags;
+
+    prstatus->regs.cs = hv->b->sregs.cs.selector;
+    prstatus->regs.ss = hv->b->sregs.ss.selector;
+    prstatus->regs.ds = hv->b->sregs.ds.selector;
+    prstatus->regs.es = hv->b->sregs.es.selector;
+    prstatus->regs.fs = hv->b->sregs.fs.selector;
+    prstatus->regs.gs = hv->b->sregs.gs.selector;
+    prstatus->regs.fs_base = hv->b->sregs.fs.base;
+    prstatus->regs.gs_base = hv->b->sregs.gs.base;
+    prstatus->pid = getpid();
+
+    /* Overwrite some register information based on
+     * the input given by the Guest */
+    struct trap_regs *regs = (struct trap_regs *)info->data;
+    prstatus->regs.rip = regs->rip;
+    prstatus->regs.cs = regs->cs;
+    prstatus->regs.eflags = regs->rflags;
+    prstatus->regs.rsp = regs->rsp;
+    prstatus->regs.ss = regs->ss;
+}
+
+
+size_t ukvm_hv_get_notes_size(int *num_notes)
+{
+    *num_notes = 1;
+    return (sizeof(Elf64_Nhdr) + 8 + sizeof(x86_elf_prstatus));
+}
+
+int ukvm_hv_dump_notes(int core_fd,
+        struct ukvm_hv *hv, struct ukvm_dump_core *info)
+{
+    int rc;
+    Elf64_Nhdr nhdr = { 0 };
+    x86_elf_prstatus prstatus = { 0 };
+    ukvm_hv_fill_elf_prstatus(&prstatus, hv, info);
+    nhdr.n_namesz = 8; // strlen("core123") + 1;
+    nhdr.n_descsz = sizeof(x86_elf_prstatus);
+    nhdr.n_type = NT_PRSTATUS;
+
+    /* Write note */
+    rc = write(core_fd, &nhdr, sizeof(nhdr));
+    rc = write(core_fd, "core123", nhdr.n_namesz);
+    rc = write(core_fd, &prstatus, sizeof(x86_elf_prstatus));
+    return rc;
 }
